move hook to reminders
