+++ client/src/lib/user/hooks/usePinCollection.ts
@@ -4,7 +4,7 @@ import utc from 'dayjs/plugin/utc';
 import useUserState from '../state/state';
 import usePinnedCollections from './usePinnedCollections';
 import * as metrics from '../../metrics';
-import useConfirmPracticeReminders from '../../sessions/hooks/useConfirmPracticeReminders';
+import useConfirmPracticeReminders from '../../reminders/hooks/useConfirmPracticeReminders';
 
 dayjs.extend(utc);
